Yet again "better" avoidance, still far from perfect

+++ trunk/ViewController/Controller/UIAssetController.cpp
@@ -415,7 +415,7 @@ void UIAssetController::onActionRelease(const float frameTime)
         if(MouseState::mouseX > pMiniMap->getPosition().x && MouseState::mouseX < (pMiniMap->getPosition().x + pMiniMap->getSize())
             && MouseState::mouseY > pMiniMap->getPosition().y && MouseState::mouseY < (pMiniMap->getPosition().y + pMiniMap->getSize()))
         {
-            D3DXVECTOR2 target = pMiniMap->minimapPositionToWorld(D3DXVECTOR2(MouseState::mouseX, MouseState::mouseY));
+            D3DXVECTOR2 target = pMiniMap->minimapPositionToWorld(D3DXVECTOR2((FLOAT)MouseState::mouseX, (FLOAT)MouseState::mouseY));
             m_pUnitCommandDispatcher->getTarget()->setTarget((const unsigned short)target.x, (const unsigned short)target.y, false);
             m_pUnitCommandDispatcher->dispatch(KeyboardState::keyDown[m_KeyQueueCommands]);
             return;
